management-complex: update SHA and checksum on LICENSE file
The MC 10.35.0 release tag was updated to a new SHA that brings some
changes in the SCR file and the LICENSE one. The MC binaries were not
touched and they are the same as before.

Update the recipe to reflect the new SHA and also the new license file
name and its md5 checksum.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>